requirements.txt
flask>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-anthropic>=0.0.1
anthropic>=0.8.0
python-docx>=0.8.11
supabase>=1.0.0
openpyxl>=3.0.0
werkzeug>=2.0.0
asgiref>=3.0.0